Support for verifying catalog signed files on windows.
On windows many operating system binaries are catalog signed,
i.e. the signature information for those binaries resides in the
windows catalog database instead of being embedded in the binaries
itself. To be able to to verify signature for such files need to
find the catalog file containing that information (based on PE file
hash) and then pass it to WinVerifyTrust to validate the signature
information.

Had to change from using winapi crate to windows_sys crate because
some of the windows apis were not available in winapi crate.